hw 2 done
