+++ controllers/index.js
@@ -1,0 +1,15 @@
+const {
+  ctrlListContact,
+  ctrlAddContact,
+  ctrlGetContactById,
+  ctrlUpdateContact,
+  ctrlRemoveContact,
+} = require("./controllers");
+
+module.exports = {
+  ctrlListContact,
+  ctrlAddContact,
+  ctrlGetContactById,
+  ctrlUpdateContact,
+  ctrlRemoveContact,
+};
