+++ models/contacts.json
@@ -58,11 +58,5 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "gmZ8x27JYS52kCAftUR8X",
-    "name": "ni",
-    "email": "nikolay123@kiev.ua",
-    "phone": "1234567890"
   }
 ]
