+++ validation/index.js
@@ -1,0 +1,6 @@
+const { validAddContact, validUpdateContact } = require("./validation");
+
+module.exports = {
+  validAddContact,
+  validUpdateContact,
+};
